The first wave of refactoring is done. Consult the docs/data_structures.txt for what would be done tomorrow.

+++ Commander/src/group2/sdp/pc/server/skeleton/ServerSkeleton.java
@@ -43,5 +43,5 @@ public interface ServerSkeleton {
 	 */
 	public void sendSpinRight(int speed, int angle);
 	
-	public void sendMoveArc(int radius, int angle);
+	public void sendMoveArc(float radius, int angle);
 }
